Move initialisation of UiConfig out of MainWindow.xaml.cs
