+++ WalletWasabi.Gui/ViewModels/MainWindowViewModel.cs
@@ -3,11 +3,14 @@ using AvalonStudio.Extensibility;
 using AvalonStudio.Extensibility.Dialogs;
 using AvalonStudio.Shell;
 using ReactiveUI;
-using System;
+using Splat;
+using System.IO;
+using System.Linq;
 using System.Reactive;
+using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using WalletWasabi.Gui.Controls.LockScreen;
-using WalletWasabi.Logging;
+using WalletWasabi.Gui.Tabs.WalletManager;
 
 namespace WalletWasabi.Gui.ViewModels
 {
@@ -69,6 +72,48 @@ namespace WalletWasabi.Gui.ViewModels
 		public MainWindowViewModel()
 		{
 			Shell = IoC.Get<IShell>();
+
+			var global = Locator.Current.GetService<Global>();
+
+			if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+			{
+				var uiConfig = global.UiConfig;
+
+				Width = uiConfig.Width;
+				Height = uiConfig.Height;
+				WindowState = uiConfig.WindowState;
+			}
+			else
+			{
+				WindowState = WindowState.Maximized;
+			}
+
+			StatusBar = new StatusBarViewModel();
+
+			LockScreen = new LockScreenViewModel();
+
+			LockScreen.Initialize();
+
+			DisplayWalletManager();
+		}
+
+		private void DisplayWalletManager()
+		{
+			var walletManagerViewModel = IoC.Get<WalletManagerViewModel>();
+			IoC.Get<IShell>().AddDocument(walletManagerViewModel);
+
+			var global = Locator.Current.GetService<Global>();
+
+			var isAnyDesktopWalletAvailable = Directory.Exists(global.WalletsDir) && Directory.EnumerateFiles(global.WalletsDir).Any();
+
+			if (isAnyDesktopWalletAvailable)
+			{
+				walletManagerViewModel.SelectLoadWallet();
+			}
+			else
+			{
+				walletManagerViewModel.SelectGenerateWallet();
+			}
 		}
 
 		public IShell Shell { get; }
